requirements.txt
requests==2.28.1
numpy==1.23.3
pandas==1.5.0
scikit-learn==1.2.1
python-telegram-bot==20.1
Flask==2.3.2
gunicorn==20.1.0
tensorflow==2.12.0